+++ dashboard/src/routes.tsx
@@ -1,7 +1,7 @@
 import App from "App";
-import Admin from "pages/Admin";
+import AdminServersPage from "pages/AdminServersPage";
 import AdminGuard from "pages/AdminGuard";
-import AdminServer from "pages/AdminServer";
+import AdminServerPage from "pages/AdminServerPage";
 import Auth from "pages/Auth";
 import AuthGuard from "pages/AuthGuard";
 import Dashboard from "pages/Dashboard";
@@ -12,6 +12,7 @@ import SettingsPage from "pages/SettingsPage";
 import SubscriptionPage from "pages/Subscription";
 import TrackPage from "pages/TrackPage";
 import { createRoutesFromElements, Navigate, Route } from "react-router-dom";
+import AdminPage from "pages/AdminPage";
 
 const routes = createRoutesFromElements(
   <>
@@ -30,9 +31,11 @@ const routes = createRoutesFromElements(
         </Route>
       </Route>
       <Route element={<AdminGuard redirectTo="/" />}>
-        <Route path="admin">
-          <Route index element={<Admin />} />
-          <Route path=":serverId" element={<AdminServer />} />
+        <Route path="admin" element={<AdminPage />}>
+          <Route index element={<Navigate to="servers" replace={true} />} />
+          <Route path="servers" element={<AdminServersPage />} />
+          <Route path="servers/:serverId" element={<AdminServerPage />} />
+          <Route path="subscriptions" element={<AdminServersPage />} />
         </Route>
       </Route>
       <Route
